Added MainForm Moved several functions from SettingsForm to Functions and MainForm
