+++ SettingsForm.cs
@@ -36,41 +36,22 @@ namespace twitch_stream_check
         /// </summary>
         public MySettings settings;
         /// <summary>
-        /// Holds the last response from a web request (make sure to not make rival requests at the same time)
-        /// </summary>
-        private HttpWebResponse HttpWResp;
-        /// <summary>
-        /// Is true if there is currently a Streamlist check running
-        /// </summary>
-        private bool bGettingData;
-        /// <summary>
-        /// The running number of the current check
-        /// </summary>
-        private int iCurrentCheck;
-        /// <summary>
-        /// Current number of streamlist entries to check
-        /// </summary>
-        private int iMaxCheck;
-        /// <summary>
-        /// Currently active streams to display in icon tooltip
-        /// </summary>
-        private int _iActiveStreams;
-        /// <summary>
-        /// Background timer to call events
-        /// </summary>
-        private System.Timers.Timer tBackgroundTimer;
-        /// <summary>
         /// Use logging for exceptions
         /// </summary>
         private Logging Log;
+        /// <summary>
+        /// Hold object of MainForm
+        /// </summary>
+        public MainForm objMainForm;
         
         
-        public SettingsForm()
+        public SettingsForm(MainForm mainform)
         {
             //
             // The InitializeComponent() call is required for Windows Forms designer support.
             //
             InitializeComponent();
+            objMainForm = mainform; // make mainform accessible from settings form
             
             // try to hide the form on startup
             this.Hide();
@@ -79,8 +60,9 @@ namespace twitch_stream_check
             this.Log = new Logging();
             
             // load stored settings from file and check if default values need to be put in
-            this.settings = MySettings.Load();
-            putDefaultSettings();
+//            this.settings = MySettings.Load(); // do this in MainForm!
+            
+            
         }
         
         public MySettings Start()
@@ -97,11 +79,6 @@ namespace twitch_stream_check
             // put supplied settings into working space
             this.settings = objSettings;
             
-            // set the check value if a check is currently running
-            bGettingData = false;
-            iCurrentCheck = 0;
-            iMaxCheck = 0;
-            iActiveStreams = 0;
             // add settings into form
             putSettingsIntoForm();
             
@@ -122,7 +99,7 @@ namespace twitch_stream_check
             linkLabelFeedback.Links.Add( i, 13, "mailto:"+sTMP+"?subject=[TwitchStreamCheckerFeedback]");
             sTMP = null; // clear tempvar
             // get 2nd occurance of Send Feedback to avoid counting it over and over again
-            i = GetNthIndex(linkLabelFeedback.Text, "Send Feedback", 2);
+            i = Functions.GetNthIndex(linkLabelFeedback.Text, "Send Feedback", 2);
             Debug.WriteLineIf(GlobalVar.DEBUG, "SETTINGSFORM: Send Feedback 2 located at: " + i);
             linkLabelFeedback.Links.Add( i, 13, "steam://friends/message/76561197960330502"); // autositz: 76561197960330502  kretze: 76561197993179564
             
@@ -135,74 +112,9 @@ namespace twitch_stream_check
 //            tTimerFirst.Enabled = false;
 //            Task.Factory.StartNew(checkStreams);
             
-            // create a new timer to run stuff at given interval
-            tBackgroundTimer = new System.Timers.Timer();
-            // FIXME RELEASE: Make sure to set checkinterval timer for release!
-            // start timer with 1000ms * 60s * interval-minutes
-            tBackgroundTimer.Interval = (settings.checkinterval * 60 * 1000);
-//            tBackgroundTimer.Interval = 5000; // uncomment this for faster cycles on small entries
-            // redo associated actions
-            tBackgroundTimer.AutoReset = true;
-            // set the action we want to do at the given interval
-            tBackgroundTimer.Elapsed += new ElapsedEventHandler(doTimer);
-            // make sure the timer is starting
-            tBackgroundTimer.Enabled = true; // enable timer
             
             
             return this.settings;
-        }
-        
-        /// <summary>
-        /// What to do each time our timer calls for us
-        /// </summary>
-        void doTimer(object sender, EventArgs e)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "DOTIMER: Timer called us");
-            Task.Factory.StartNew(checkStreams);
-            
-        }
-        
-        /// <summary>
-        /// Close the program
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="e"></param>
-        void ToolStripMenuItemClickExit(object sender, EventArgs e)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "TOOLSTRIPMENUITEMCLICKEXIT: Menu Click - Exit");
-            if (HttpWResp != null) {
-                HttpWResp.Close();
-                Debug.WriteLineIf(GlobalVar.DEBUG, "TOOLSTRIPMENUITEMCLICKEXIT: WebResponse closed");
-            }
-                
-            Application.Exit();
-        }
-        
-        /// <summary>
-        /// About info
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="e"></param>
-        void ToolStripMenuItemClickAbout(object sender, EventArgs e)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "TOOLSTRIPMENUITEMCLICKABOUT: Menu Click - About");
-            MessageBox.Show("This tool will check if any of the configured streams is currently online.", "Twitch Stream Checker");
-        }
-        
-        /// <summary>
-        /// Open the Settings menu and populate it with current settings data
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="e"></param>
-        void ToolStripMenuItemClickSettings(object sender, EventArgs e)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "TOOLSTRIPMENUITEMCLICKSETTINGS: Menu Click - Settings");
-            // populate settings form with currently store data
-            putSettingsIntoForm();
-            
-            // make sure we see a settings window which was previously hidden, hopefully...
-            this.Show();
-            this.Focus();
         }
         
         /// <summary>
@@ -223,21 +135,25 @@ namespace twitch_stream_check
             Debug.WriteLineIf(GlobalVar.DEBUG, "BUTTONSETTINGSCLICKSAVE: Store the form field data into the settings");
             // get current settings from form and store them in the file
             // check interval
-            this.settings.checkinterval = convertNum(comboBoxInterval.Text, 10);
-            comboBoxInterval.Text = this.settings.checkinterval.ToString();
-            tBackgroundTimer.Interval = this.settings.checkinterval * 60 * 1000;
-            Debug.WriteLineIf(GlobalVar.DEBUG, "BUTTONSETTINGSCLICKSAVE: Set new interval to: " + tBackgroundTimer.Interval + " ms");
+            int iTMP = Functions.convertNum(comboBoxInterval.Text, 10);
+            // set timer to the new value entered in settings
+            if (iTMP != this.settings.checkinterval) {
+                this.settings.checkinterval = iTMP;
+                objMainForm.SetTimer(this.settings.checkinterval);
+            }
             
             // check account name
-            this.settings.checkaccount = convertAlphaNum(textBoxAccountCheck.Text);
-            textBoxAccountCheck.Text = this.settings.checkaccount;
+            this.settings.checkaccount = Functions.convertAlphaNum(textBoxAccountCheck.Text);
             Debug.WriteLineIf(GlobalVar.DEBUG, "BUTTONSETTINGSCLICKSAVE: Set new Account to: " + textBoxAccountCheck.Text);
             
             // check streams
-            
+            // they are stored directly in settings.streams
             
             this.settings.Save();
             bSender.Enabled = true;
+            // TODO MAIN: make sure all data is stored in mainform!!!
+            objMainForm.settings = this.settings;
+            this.Dispose();
         }
         /// <summary>
         /// Close the Settings menu without storing data
@@ -248,6 +164,31 @@ namespace twitch_stream_check
         {
             Debug.WriteLineIf(GlobalVar.DEBUG, "BUTTONSETTINGSCLICKCANCEL: Button click - Cancel");
             this.Visible = false;
+            // TODO MAIN: make sure all data is stored in mainform!!!
+            this.Dispose();
+        }
+        
+        /// <summary>
+        /// Close the Settings menu without storing data
+        /// </summary>
+        /// <param name="sender">SettingsForm</param>
+        /// <param name="e">FormClosingEventArgs</param>
+        private void SettingsForm_FormClosing(object sender, FormClosingEventArgs e)
+        {
+            Debug.WriteLineIf(GlobalVar.DEBUG, "SETTINGSFORM_FORMCLOSING: Someone clicked Close");
+            if (e.CloseReason == CloseReason.UserClosing) {
+                Debug.WriteLineIf(GlobalVar.DEBUG, "SETTINGSFORM_FORMCLOSING: User closed the form");
+                e.Cancel = true;
+                
+                // TODO: lazy work to hide the form, remove it and recreate maybe? no need to hold the settings form in memory all the time i guess...
+                SettingsForm senders = sender as SettingsForm;
+                if (senders != null) {
+                    senders.Hide();
+                }
+                
+                // TODO MAIN: make sure all data is stored in mainform!!!
+                this.Dispose();
+            }
         }
         
         /// <summary>
@@ -265,74 +206,13 @@ namespace twitch_stream_check
         }
         
         /// <summary>
-        /// Convert the settings dialog data to settings save data.
-        /// </summary>
-        /// <param name="s">String with linefeeds to be converted</param>
-        /// <returns>String Array</returns>
-        public String[] convertLFtoArray(String s)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CONVERTLFTOARRAY: Convert a string with newlines to an array that splits at newlines");
-            return s.Split(new string[] { "\r\n", "\n", "\r" }, StringSplitOptions.None);
-        }
-        
-        /// <summary>
-        /// Convert the settings savefile data to settings dialog output.
-        /// </summary>
-        /// <param name="a">String Array to be merged</param>
-        /// <returns>String with line feeds</returns>
-        public String convertArraytoLF(String[] a)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CONVERTARRAYTOLF: Convert an array to string with newlines");
-            return string.Join(Environment.NewLine, a);
-        }
-        
-        /// <summary>
-        /// Clear any non-alphanumeric from text with the option to clear or preserve newlines.
-        /// </summary>
-        /// <param name="s">String to be checked</param>
-        /// <param name="preserveNewline">Keep new lines or strip them too</param>
-        /// <returns>Cleaned string</returns>
-        public string convertAlphaNum(string s, bool preserveNewline= false)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CONVERTALPHANUM: Converting text string to contain only valid chars with by preserving newline: " + preserveNewline);
-            char[] arr = s.ToCharArray();
-            
-            if (preserveNewline) {
-                arr = Array.FindAll<char>(arr, (c => (char.IsLetterOrDigit(c) || c == '\n' || c == '\r' || c == '_')));
-            } else {
-                arr = Array.FindAll<char>(arr, (c => (char.IsLetterOrDigit(c) || c == '_')));
-            }
-            
-            String str = new string(arr);
-            
-            return str;
-        }
-        
-        /// <summary>
-        /// Clear any non-numeric from text with fallback integer value
-        /// </summary>
-        /// <param name="s">String to be converted to an int</param>
-        /// <param name="iFallback">Integer value to be used when conversion fails</param>
-        /// <returns>Converted Integer or Fallback value</returns>
-        public int convertNum(string s, int iFallback = 0)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CONVERTNUM: Converting " + s + " with fallback: " + iFallback);
-            int i = 0;
-            if (!Int32.TryParse(s, out i)) {
-                i = iFallback;
-            }
-            
-            return i;
-        }
-        
-        /// <summary>
         /// Put currently stored settings into settings form
         /// </summary>
         private void putSettingsIntoForm()
         {
             Debug.WriteLineIf(GlobalVar.DEBUG, "PUTSETTINGSINTOFORM: Get settings into the form fields");
-            comboBoxInterval.Text = convertNum(settings.checkinterval.ToString()).ToString();
-            textBoxAccountCheck.Text = convertAlphaNum(this.settings.checkaccount);
+            comboBoxInterval.Text = Functions.convertNum(settings.checkinterval.ToString()).ToString();
+            textBoxAccountCheck.Text = Functions.convertAlphaNum(this.settings.checkaccount);
             
             try {
                 
@@ -377,15 +257,17 @@ namespace twitch_stream_check
         {
             Debug.WriteLineIf(GlobalVar.DEBUG, "GETACCOUNTFOLLOWINGS: Parsing done");
             bool bSuccess = false;
-            sAccount = convertAlphaNum(sAccount);
+            WebCheckResponse WebCheck;
+            sAccount = Functions.convertAlphaNum(sAccount);
             string sBaseURL = "https://api.twitch.tv/kraken/streams/followed/";
             string sParams = "";
             string responseString = "";
             
             string sURL = sBaseURL + sAccount + sParams;
             
-            if (doWebRequest(sURL)) {
-                using (Stream stream = HttpWResp.GetResponseStream()) {
+            WebCheck = Functions.doWebRequest(sURL);
+            if (WebCheck.HttpWResp.StatusCode == HttpStatusCode.OK) {
+                using (Stream stream = WebCheck.HttpWResp.GetResponseStream()) {
                     StreamReader reader = new StreamReader(stream, Encoding.UTF8);
                     responseString = reader.ReadToEnd();
                 }
@@ -399,328 +281,6 @@ namespace twitch_stream_check
             return bSuccess;
         }
         
-        /// <summary>
-        /// Check if an account is currently online
-        /// </summary>
-        /// <param name="sAccount">Streamname</param>
-        /// <returns>Returns game name if stream is online</returns>
-        public string getOnlineStatus(string sAccount)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: Putting response into var");
-            string sReturn = null;
-            sAccount = convertAlphaNum(sAccount);
-            const string sBaseURL = "https://api.twitch.tv/kraken/streams/";
-            string sParams = "";
-            string responseString = "";
-            
-            string sURL = sBaseURL + sAccount + sParams;
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: Created URL: " + sURL);
-            if (sAccount != "" && doWebRequest(sURL)) {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: Parsing WebResponse");
-                try {
-                    using (Stream stream = HttpWResp.GetResponseStream()) {
-                        StreamReader reader = new StreamReader(stream, Encoding.UTF8);
-                        responseString = reader.ReadToEnd();
-                    }
-                } catch (Exception ex) {
-                    Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: Failed on datastream: " + ex.Message);
-                    Log.Add("getOnlineStatus>" + ex.Message);
-                }
-                
-                
-                Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: Creating workable object from json response");
-                try {
-                    dynamic data = JsonConvert.DeserializeObject(responseString);
-                    if (data.stream == null) {
-                        Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: data.stream is null - user offline or error");
-                        // offline handling
-                        
-                    } else {
-                        Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: data.stream contains something - user should be online");
-                        // online handling
-                        
-                        sReturn = data.stream.game;
-                    }
-                } catch (Exception ex) {
-                    Debug.WriteLineIf(GlobalVar.DEBUG, "!EXCEPTION!:GETONLINESTATUS: Failed on Deserialize: " + ex.Message);
-                    Log.Add("getOnlineStatus>" + ex.Message);
-                }
-                
-                
-                
-            }
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETONLINESTATUS: User is playing: " + sReturn);
-            return sReturn;
-        }
-        
-        /// <summary>
-        /// Get data from an URL
-        /// </summary>
-        /// <param name="sURL">URL to contact</param>
-        /// <returns>Returns true if request was successfull</returns>
-        public bool doWebRequest(string sURL)
-        {
-            bool bRet = false;
-            
-            try {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "DOWEBREQUEST: Making a new Web Request");
-                HttpWebRequest HttpWReq = (HttpWebRequest)WebRequest.Create(sURL);
-                Debug.WriteLineIf(GlobalVar.DEBUG, "DOWEBREQUEST: Putting response into var");
-                HttpWResp = (HttpWebResponse)HttpWReq.GetResponse();
-                Debug.WriteLineIf(GlobalVar.DEBUG, "DOWEBREQUEST: Response stored for future use");
-                
-                if (HttpWResp.StatusCode == HttpStatusCode.OK)
-                    bRet = true;
-            } catch (Exception ex) {
-                Log.Add("doWebRequest>" + ex.Message);
-            }
-            
-            return bRet;
-        }
-        
-        /// <summary>
-        /// Create a new menu entry to follow the stream, but check if it already exists
-        /// </summary>
-        /// <param name="sUser">Streamname</param>
-        private bool MenuEntryCreate(string sUser, string sGame)
-        {
-            // make sure we can access the settingsform
-            if (this.InvokeRequired)
-            {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: NEEDS INVOKE");
-                return (bool)this.Invoke ((Func<string,string,bool>)MenuEntryCreate, sUser, sGame);
-            }
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: START");
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: Creating new Menu entry for " + sUser + " playing game " + sGame);
-            bool bRet = false;
-            int iMenuIDX = MyMenu.Items.IndexOfKey(sUser);
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: Result of looking up menu entry: " + iMenuIDX);
-            
-            // -1 means we have no matching menu entry and will add a new one, otherwise do nothing
-            if (iMenuIDX == -1) {
-                notifyIcon1.ShowBalloonTip(900, "Stream is LIVE", sUser + " playing " + sGame, ToolTipIcon.None);
-
-                // add the new stream on top of the menu and append the other entries back to the menu
-                // make sure to add name to be able to search by key!
-                ToolStripItem tsiNewItem = new ToolStripMenuItem(sUser, null, (sender, e) => openStream(sUser), sUser);
-                tsiNewItem.ToolTipText = "Playing: " + convertAlphaNum(sGame);
-                Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: Created new MenuItem");
-                // add the new stream on top of the menu
-                try {
-                    MyMenu.Invoke((MethodInvoker) delegate {
-                                      lock (MyMenu) {
-                                          MyMenu.Items.Insert(0, tsiNewItem);
-                                      }
-                                  });
-//                    MyMenu.Items.Insert(0, tsiNewItem);
-                    bRet = true;
-                    // increase active streams number only when we are really adding a new one
-                    iActiveStreams++;
-                } catch (Exception ex) {
-                    object[] tmp = ParseException(ex);
-                    Debug.WriteLineIf(GlobalVar.DEBUG, "!EXCEPTION!:MENUENTRYCREATE: Error inserting menu item (" + tmp[2] + ":" + tmp[3] + ":"  + tmp[0] + "): " + ex.Message);
-                    Log.Add("MenuEntryCreate>" + ex.Message);
-                }
-                
-                Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: Item added to Menu");
-            }
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYCREATE: END");
-            return bRet;
-        }
-        
-        /// <summary>
-        /// Check if a menu entry exists and remove it
-        /// </summary>
-        /// <param name="sUser">Streamname</param>
-        private bool MenuEntryRemove(string sUser)
-        {
-            // make sure we can access the settingsform
-            if (this.InvokeRequired)
-            {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYREMOVE: NEEDS INVOKE");
-                return (bool)this.Invoke ((Func<string,bool>)MenuEntryRemove, sUser);
-            }
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYREMOVE: Remove menu entry for stream: " + sUser);
-            bool bRet = false;
-            MyMenu.Invoke((MethodInvoker) delegate {
-                              lock (MyMenu) {
-                                  int iMenuIDX = MyMenu.Items.IndexOfKey(sUser);
-                                  Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYREMOVE: Menu index: " + iMenuIDX);
-                                  
-                                  
-                                  if (iMenuIDX != -1) {
-                                      Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYREMOVE: Entry found at: " + iMenuIDX);
-                                      notifyIcon1.ShowBalloonTip(750, "Stream is Offline", sUser, ToolTipIcon.Info);
-                                                            MyMenu.Items.RemoveAt(iMenuIDX);
-//                                      MyMenu.Items.RemoveAt(iMenuIDX);
-                                      bRet = true;
-                                      // decrease active streams number only when we are really removing an entry
-                                      iActiveStreams--;
-                                  }
-                              }
-                          });
-            
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "MENUENTRYREMOVE: END");
-            return bRet;
-        }
-        
-        private bool updateToolTip()
-        {
-            if (this.InvokeRequired)
-                return (bool)this.Invoke ((Func<bool>)updateToolTip);
-            
-            string s = "";
-            // display info when we are currently running an update
-            if (bGettingData) {
-                s = iActiveStreams + " active Streams (Updating " + iCurrentCheck + "/" + iMaxCheck + ")";
-            } else {
-                s = iActiveStreams + " active Streams";
-            }
-            notifyIcon1.Text = s;
-//            MyMenu.Invoke((MethodInvoker) delegate {notifyIcon1.Text = s;});
-            
-            return true;
-        }
-        
-        /// <summary>
-        /// Open stream in default browser
-        /// </summary>
-        /// <param name="sUser">Streamname</param>
-        public void openStream(string sUser)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "OPENSTREAM: Launch a browser to watch stream");
-            string sURL = getStreamURL(convertAlphaNum(sUser));
-            Debug.WriteLineIf(GlobalVar.DEBUG, "OPENSTREAM: Opening URL: " + sURL);
-            try {
-                Process.Start(sURL);
-            } catch (Exception ex) {
-                Log.Add("openStream>" + ex.Message);
-            }
-        }
-        
-        /// <summary>
-        /// Creates an URL to a supplied Twitch Stream
-        /// </summary>
-        /// <param name="sUser">Streamname</param>
-        /// <returns>Returns URL of a twitch stream</returns>
-        public string getStreamURL(string sUser)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETSTREAMURL: Create an URL for browser");
-            const string sBaseURL   = "http://twitch.tv/";
-            const string sAppendURL = "/";
-            
-            // create new url and make sure there is no sneaky code in the username
-            string sURL = sBaseURL + convertAlphaNum(sUser) + sAppendURL;
-            
-            return sURL;
-        }
-        
-        /// <summary>
-        /// Get information if a user is currently streaming
-        /// </summary>
-        /// <param name="sUser">Streamname</param>
-        public bool checkUser(string sUser)
-        {
-            bool bActive = false;
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CHECKUSER: We lookup a user: " + sUser);
-            //notifyIcon1.ShowBalloonTip(1500, "Checking user", sUser, ToolTipIcon.Info);
-            string sGame = getOnlineStatus(sUser);
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CHECKUSER: We lookuped a user");
-#if _HACK
-            // HACK: creating random entries
-            if (GlobalVar.GENERATEDATA) {
-                Random rnd = new Random();
-                char[] let = new char[10];
-                for (int i = 0; i < 10; i++) {
-                    let[i] = (char)('a' + rnd.Next(0, 26));
-                }
-                sGame = new string(let);
-                if (rnd.Next(0, 1000) < 500)
-                    sGame = null;
-                Debug.WriteLineIf(GlobalVar.DEBUG, "HACK:CHECKUSER: Random user: " + sGame);
-            }
-#endif
-            
-            if (sGame != null) {
-                //MessageBox.Show(sUser + Environment.NewLine + sGame, "Creating Menu");
-                MenuEntryCreate(sUser, sGame);
-                bActive = true;
-            } else {
-                MenuEntryRemove(sUser);
-            }
-            
-            return bActive;
-        }
-        
-        /// <summary>
-        /// Check all configured streams
-        /// </summary>
-        public void checkStreams()
-        {
-            // create a local copy of the current list of streams to avoid getting a mixup when changing settings, settings were used directly earlier
-            IList<twStream> aCheckUsers = settings.streams; // store a local copy of streamlist
-            
-            Debug.WriteLineIf(GlobalVar.DEBUG, "CHECKSTREAMS: We gotta check our list of streams");
-            if (!bGettingData) {
-                bGettingData = true;
-                
-                // set maxcheck for progress display
-                iMaxCheck = aCheckUsers.Count;
-                try {
-                    for (int i = 0; i < iMaxCheck; i++) {
-                        iCurrentCheck = i + 1;
-                        checkUser(aCheckUsers[i].sStreamname);
-                    }
-                } catch (Exception ex) {
-                    // user index could be out of bounds
-                    Log.Add("checkStreams>" + ex.Message);
-                }
-                
-                bGettingData = false;
-            } else {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "CHECKSTREAMS: Another check is still running..." + iCurrentCheck + "/" + iMaxCheck);
-                // FIXME RELEASE: Make sure to have this enabled for release! So users get a warning when their list can't be processed in between intervals.
-                MessageBox.Show("Consider raising the interval a bit." + Environment.NewLine + "Processing " + iCurrentCheck + "/" + iMaxCheck + Environment.NewLine + Environment.NewLine +
-                                "If this message keeps coming up over and over again at the same step then consider restarting the program.",
-                                "Check already running", MessageBoxButtons.OK, MessageBoxIcon.Exclamation, MessageBoxDefaultButton.Button1);
-            }
-        }
-        
-        /// <summary>
-        /// Tries to get more detailed info on the exception
-        /// </summary>
-        /// <param name="e">Supply the Exception that has occured.</param>
-        /// <returns>Returns object: 0 Line number, 1 Column number, 2 Filename, 3 Methodname</returns>
-        public object[] ParseException(Exception ex)
-        {
-            object[] data = { 0, 0, "", "" }; // <int>line, <int>column, <string>filename, <string>methodname
-            
-            //Get a StackTrace object for the exception
-            StackTrace st = new StackTrace(ex, true);
-        
-            //Get the first stack frame
-            StackFrame frame = st.GetFrame(0);
-        
-            //Get the file name
-            data[2] = frame.GetFileName();
-        
-            //Get the method name
-            data[3] = frame.GetMethod().Name;
-        
-            //Get the line number from the stack frame
-            data[0] = frame.GetFileLineNumber();
-        
-            //Get the column number
-            data[1] = frame.GetFileColumnNumber();
-            
-            return data;
-        }
-        
         private void linkLabelFeedback_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
         {
             string sLink = e.Link.LinkData.ToString();
@@ -729,22 +289,6 @@ namespace twitch_stream_check
                 Process.Start(sLink);
             } catch (Exception ex) {
                 Log.Add("linkLabelFeedback_LinkClicked>" + ex.Message);
-            }
-        }
-        
-        private void SettingsForm_FormClosing(object sender, FormClosingEventArgs e)
-        {
-            Debug.WriteLineIf(GlobalVar.DEBUG, "SETTINGSFORM_FORMCLOSING: Someone clicked Close");
-            if (e.CloseReason == CloseReason.UserClosing) {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "SETTINGSFORM_FORMCLOSING: User closed the form");
-                e.Cancel = true;
-                
-                // TODO: lazy work to hide the form, remove it and recreate maybe? no need to hold the settings form in memory all the time i guess...
-                SettingsForm senders = sender as SettingsForm;
-                if (senders != null) {
-                    senders.Hide();
-                }
-                
             }
         }
         
@@ -800,17 +344,6 @@ namespace twitch_stream_check
         }
         
         /// <summary>
-        /// Wrapper to update ToolTip everytime the streamer count changes
-        /// </summary>
-        public int iActiveStreams {
-            get{ return _iActiveStreams; }
-            set {
-                _iActiveStreams = value;
-                updateToolTip();
-            }
-        }
-        
-        /// <summary>
         /// Handle clicks into the Streamlist cells
         /// </summary>
         /// <param name="sender">DataGridView</param>
@@ -829,72 +362,6 @@ namespace twitch_stream_check
                     Log.Add("dgvStreams_CellContentClick>" + ex.Message);
                 }
             }
-        }
-        
-        /// <summary>
-        /// Enter default settings values if non have been loaded
-        /// </summary>
-        public void putDefaultSettings()
-        {
-            if (settings.checkinterval == 0 || settings.checkinterval < 1) {
-                settings.checkinterval = 5;
-            }
-            
-            if (settings.checkaccount == null) {
-                settings.checkaccount = "Account";
-            }
-            
-            // add default entries when list is empty
-            if (settings.streams == null) {
-                // FIXME RELEASE: set default streams
-                string[] streamers = new string[] {"Autositz", "DRUCKWELLETV", "Garrynewman", "Denkii"};
-//                string[] streamers = new string[] {"Entry 1", "Entry 2", "Entry 3", "Entry 4"};
-                settings.streams = new List<twStream>();
-                foreach (string s in streamers) {
-                    twStream s2 = new twStream();
-                    s2.bImportant= true;
-                    s2.sStreamname = s;
-                    settings.streams.Add(s2);
-                }
-            }
-            
-        }
-        
-        public int GetNthIndex(string sText, string sSearch, int iOccurred)
-        {
-            return GetNthIndex(sText, sSearch, iOccurred, 0, 1);
-        }
-        /// <summary>
-        /// Finds the position of the Nth occurrance of Search in Text
-        /// </summary>
-        /// <param name="sText"></param>
-        /// <param name="sSearch"></param>
-        /// <param name="iOccurred">int </param>
-        /// <param name="iPos"></param>
-        /// <param name="iRunning"></param>
-        /// <returns>Position of the Nth Occurrance</returns>
-        public int GetNthIndex(string sText, string sSearch, int iOccurred, int iPos, int iRunning)
-        {
-            int iRet = -1;
-            // search index, increase starting point and interval if something was found
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETNTHINDEX: Starting search at \""+iPos+"\" for \""+sSearch+"\" in \""+sText+"\"");
-            iPos = sText.IndexOf(sSearch, iPos);
-            Debug.WriteLineIf(GlobalVar.DEBUG, "GETNTHINDEX: Found entry at: " + iPos);
-            // we are in the depth we need
-            if (iRunning == iOccurred)
-            {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "GETNTHINDEX: We found the "+iOccurred+" occurance at: "+iPos);
-                iRet = iPos;
-            } else
-            {
-                Debug.WriteLineIf(GlobalVar.DEBUG, "GETNTHINDEX: Doing another round: " + iRunning);
-                iRunning++; // increase by one to know how often we have run
-                iPos++; // increase by one to find the same entry again
-                iRet = GetNthIndex(sText, sSearch, iOccurred, iPos, iRunning);
-            }
-            
-            
-            return iRet;
         }
     }
